convert twig filters and functions into autoloadable twig extensions
