--- app/src/View/Filters.php
@@ -1,56 +1,0 @@
-<?php
-namespace View\Filters;
-
-use Twig_Environment;
-use Twig_SimpleFilter;
-
-function initialize(Twig_Environment $env)
-{
-    $env->addFilter(
-        new Twig_SimpleFilter(
-            'img_path',
-            '\View\Filters\img_path',
-            ['needs_environment' => true]
-        )
-    );
-    $env->addFilter(
-        new Twig_SimpleFilter(
-            'link',
-            '\View\Filters\link',
-            ['is_safe' => ['html']]
-        )
-    );
-    $env->addFilter(
-        new Twig_SimpleFilter(
-            'format_date',
-            '\View\Filters\format_date'
-        )
-    );
-}
-
-function img_path($env, $suffix, $infix)
-{
-    if (!$suffix && $infix = 'event_icons') {
-        $suffix = 'none.png';
-    }
-
-    $path = '/img/' . $infix . '/' . $suffix;
-
-    // Allow for migration to local images
-    if (file_exists($_SERVER['DOCUMENT_ROOT'] . $path)) {
-        $uri = $env->getExtension('slim')->base();
-        return $uri . $path;
-    }
-
-    return 'https://joind.in/inc' .$path;
-}
-
-function format_date($date)
-{
-    return date('D M dS Y', strtotime($date));
-}
-
-function link($url, $label = '', $class = '')
-{
-    return '<a href="' . $url . '" class="' . $class . '">' . ($label ? $label : $url) . '</a>';
-}
